Move endpoint stats
