+++ profiling/src/internal/endpoint_stats.rs
