+++ profiling/src/internal/endpoints.rs
@@ -1,7 +1,6 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2023-Present Datadog, Inc.
 use super::*;
-use crate::profile::profiled_endpoints::ProfiledEndpointsStats;
 
 pub struct Endpoints {
     pub endpoint_label: StringId,
