+++ profiling/src/internal/profile.rs
@@ -5,8 +5,8 @@ use self::api::UpscalingInfo;
 use super::*;
 use crate::api;
 use crate::collections::identifiable::*;
+use crate::internal::ProfiledEndpointsStats;
 use crate::pprof::sliced_proto::*;
-use crate::profile::profiled_endpoints::ProfiledEndpointsStats;
 use crate::serializer::CompressedProtobufSerializer;
 use std::borrow::Cow;
 use std::collections::HashMap;
